Add caching to circumvent ISO re-downloads (currently Ubuntu-only)
After a perfectly good download, why re-download? This pain point
comes up fast and hard once you can create multiple VMs from the same
release/edition.

cached_web_get handles the job of wget, but keeps a cache of
downloaded objects and hard links the ISO image after download so that
it is also present in the expected VM_PATH. As long as everybody is in
the same UNIX filesystem (an apparent quickget assumption, at least at
present), this is a simple and reasonable thing to do: It saves much
time *and* disk space.

The downloaded objects must be uniquely named, but the hash is still
checked after return to get_ubuntu, so hopefully it won't cause harm
when new/different ISOs show up. Time will tell with this
experiment. This function would need some redesign to support multiple
OSes, but ... one step at a time.